Fix item save - was removing unrelated intel

+++ World/Types/Item.py
@@ -49,20 +49,21 @@ class Item(BaseElement, Worthy, Named):
             npc.save()
 
     def save(self, force_insert=False, only=None):
-        for dirty in self.dirty_fields:
-            if isinstance(dirty, ForeignKeyField) \
-                    and dirty.rel_model == Place:
-                try:
-                    old_place = Item.get_by_id(self.id).place
-                except:
-                    old_place = None
-                if old_place != self.place:
-                    # remove npc_place knowledge from every player's knowledge book
-                    from World.Types.Intel import Intel
-                    from World.Types.BridgeModels import PlayerKnowledgeBook
-                    results = PlayerKnowledgeBook.select().join(Intel).where(Intel.item_place == self.id)
-                    for res in results:
-                        res.delete_instance()
+        if self.id is not None:
+            for dirty in self.dirty_fields:
+                if isinstance(dirty, ForeignKeyField) \
+                        and dirty.rel_model == Place:
+                    try:
+                        old_place = Item.get_by_id(self.id).place
+                    except:
+                        old_place = None
+                    if old_place != self.place:
+                        # remove item_place knowledge from every player's knowledge book
+                        from World.Types.Intel import Intel
+                        from World.Types.BridgeModels import PlayerKnowledgeBook
+                        results = PlayerKnowledgeBook.select().join(Intel).where(Intel.item_place == self.id)
+                        for res in results:
+                            res.delete_instance()
         super(Item, self).save()
